Refactor volume validatons into common and elasticsearch
