+++ pkg/controller/common/volume/validations/validations_test.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License 2.0;
 // you may not use this file except in compliance with the Elastic License 2.0.
 
-package volume
+package validations
 
 import (
 	"context"
@@ -217,7 +217,6 @@ func Test_getDefaultStorageClass(t *testing.T) {
 		})
 	}
 }
-
 
 func Test_getStorageClass(t *testing.T) {
 	tests := []struct {
