+++ pkg/controller/elasticsearch/validation/validations.go
@@ -34,7 +34,6 @@ const (
 	nodeRolesInOldVersionMsg               = "node.roles setting is not available in this version of Elasticsearch"
 	parseStoredVersionErrMsg               = "Cannot parse current Elasticsearch version. String format must be {major}.{minor}.{patch}[-{label}]"
 	parseVersionErrMsg                     = "Cannot parse Elasticsearch version. String format must be {major}.{minor}.{patch}[-{label}]"
-	pvcImmutableErrMsg                     = "volume claim templates can only have their storage requests increased, if the storage class allows volume expansion. Any other change is forbidden"
 	pvcNotMountedErrMsg                    = "volume claim declared but volume not mounted in any container. Note that the Elasticsearch data volume should be named 'elasticsearch-data'"
 	unsupportedConfigErrMsg                = "Configuration setting is reserved for internal use. User-configured use is unsupported"
 	unsupportedUpgradeMsg                  = "Unsupported version upgrade path. Check the Elasticsearch documentation for supported upgrade paths."
