+++ pkg/controller/elasticsearch/validation/volume_validation_test.go
@@ -60,310 +60,6 @@ func withStorageReq(claim corev1.PersistentVolumeClaim, size string) corev1.Pers
 	return *c
 }
 
-func Test_ensureClaimSupportsExpansion(t *testing.T) {
-	tests := []struct {
-		name                string
-		k8sClient           k8s.Client
-		claim               corev1.PersistentVolumeClaim
-		validateStoragClass bool
-		wantErr             bool
-	}{
-		{
-			name:                "specified storage class supports volume expansion",
-			k8sClient:           k8s.NewFakeClient(withVolumeExpansion(sampleStorageClass)),
-			claim:               sampleClaim,
-			validateStoragClass: true,
-			wantErr:             false,
-		},
-		{
-			name:                "specified storage class does not support volume expansion",
-			k8sClient:           k8s.NewFakeClient(&sampleStorageClass),
-			claim:               sampleClaim,
-			validateStoragClass: true,
-			wantErr:             true,
-		},
-		{
-			name:                "default storage class supports volume expansion",
-			k8sClient:           k8s.NewFakeClient(withVolumeExpansion(defaultStorageClass)),
-			claim:               corev1.PersistentVolumeClaim{},
-			validateStoragClass: true,
-			wantErr:             false,
-		},
-		{
-			name:                "default storage class does not support volume expansion",
-			k8sClient:           k8s.NewFakeClient(&defaultStorageClass),
-			claim:               corev1.PersistentVolumeClaim{},
-			validateStoragClass: true,
-			wantErr:             true,
-		},
-		{
-			name:                "storage class validation disabled: no-op",
-			k8sClient:           k8s.NewFakeClient(&sampleStorageClass), // would otherwise be refused: no expansion
-			claim:               sampleClaim,
-			validateStoragClass: false,
-			wantErr:             false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := EnsureClaimSupportsExpansion(context.Background(), tt.k8sClient, tt.claim, tt.validateStoragClass); (err != nil) != tt.wantErr {
-				t.Errorf("ensureClaimSupportsExpansion() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func Test_allowsVolumeExpansion(t *testing.T) {
-	tests := []struct {
-		name string
-		sc   storagev1.StorageClass
-		want bool
-	}{
-		{
-			name: "allow volume expansion: true",
-			sc:   storagev1.StorageClass{AllowVolumeExpansion: ptr.To[bool](true)},
-			want: true,
-		},
-		{
-			name: "allow volume expansion: false",
-			sc:   storagev1.StorageClass{AllowVolumeExpansion: ptr.To[bool](false)},
-			want: false,
-		},
-		{
-			name: "allow volume expansion: nil",
-			sc:   storagev1.StorageClass{AllowVolumeExpansion: nil},
-			want: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := allowsVolumeExpansion(tt.sc); got != tt.want {
-				t.Errorf("allowsVolumeExpansion() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_isDefaultStorageClass(t *testing.T) {
-	tests := []struct {
-		name string
-		sc   storagev1.StorageClass
-		want bool
-	}{
-		{
-			name: "annotated as default",
-			sc:   defaultStorageClass,
-			want: true,
-		},
-		{
-			name: "annotated as default (beta)",
-			sc:   defaultBetaStorageClass,
-			want: true,
-		},
-		{
-			name: "annotated as default (+ beta)",
-			sc: storagev1.StorageClass{ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-				"storageclass.kubernetes.io/is-default-class":      "true",
-				"storageclass.beta.kubernetes.io/is-default-class": "true",
-			}}},
-			want: true,
-		},
-		{
-			name: "no annotations",
-			sc:   storagev1.StorageClass{ObjectMeta: metav1.ObjectMeta{Annotations: nil}},
-			want: false,
-		},
-		{
-			name: "not annotated as default",
-			sc:   sampleStorageClass,
-			want: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := isDefaultStorageClass(tt.sc); got != tt.want {
-				t.Errorf("isDefaultStorageClass() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_getDefaultStorageClass(t *testing.T) {
-	tests := []struct {
-		name      string
-		k8sClient k8s.Client
-		want      storagev1.StorageClass
-		wantErr   bool
-	}{
-		{
-			name:      "return the default storage class",
-			k8sClient: k8s.NewFakeClient(&sampleStorageClass, &defaultStorageClass),
-			want:      defaultStorageClass,
-		},
-		{
-			name:      "default storage class not found",
-			k8sClient: k8s.NewFakeClient(&sampleStorageClass),
-			want:      storagev1.StorageClass{},
-			wantErr:   true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := getDefaultStorageClass(tt.k8sClient)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("getDefaultStorageClass() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("getDefaultStorageClass() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_getStorageClass(t *testing.T) {
-	tests := []struct {
-		name      string
-		k8sClient k8s.Client
-		claim     corev1.PersistentVolumeClaim
-		want      storagev1.StorageClass
-		wantErr   bool
-	}{
-		{
-			name:      "return the specified storage class",
-			k8sClient: k8s.NewFakeClient(&sampleStorageClass, &defaultStorageClass),
-			claim:     corev1.PersistentVolumeClaim{Spec: corev1.PersistentVolumeClaimSpec{StorageClassName: ptr.To[string](sampleStorageClass.Name)}},
-			want:      sampleStorageClass,
-			wantErr:   false,
-		},
-		{
-			name:      "error out if not found",
-			k8sClient: k8s.NewFakeClient(&defaultStorageClass),
-			claim:     corev1.PersistentVolumeClaim{Spec: corev1.PersistentVolumeClaimSpec{StorageClassName: ptr.To[string](sampleStorageClass.Name)}},
-			want:      storagev1.StorageClass{},
-			wantErr:   true,
-		},
-		{
-			name:      "fallback to the default storage class if unspecified",
-			k8sClient: k8s.NewFakeClient(&sampleStorageClass, &defaultStorageClass),
-			claim:     corev1.PersistentVolumeClaim{Spec: corev1.PersistentVolumeClaimSpec{}},
-			want:      defaultStorageClass,
-			wantErr:   false,
-		},
-		{
-			name:      "error out if unspecified and default storage class not found",
-			k8sClient: k8s.NewFakeClient(&sampleStorageClass),
-			claim:     corev1.PersistentVolumeClaim{Spec: corev1.PersistentVolumeClaimSpec{}},
-			want:      storagev1.StorageClass{},
-			wantErr:   true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := getStorageClass(tt.k8sClient, tt.claim)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("getStorageClass() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !comparison.Equal(&got, &tt.want) {
-				t.Errorf("getStorageClass() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestValidateClaimsUpdate(t *testing.T) {
-	type args struct {
-		k8sClient            k8s.Client
-		initial              []corev1.PersistentVolumeClaim
-		updated              []corev1.PersistentVolumeClaim
-		validateStorageClass bool
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "same claims: ok",
-			args: args{
-				k8sClient:            k8s.NewFakeClient(withVolumeExpansion(sampleStorageClass)),
-				initial:              []corev1.PersistentVolumeClaim{sampleClaim, sampleClaim2},
-				updated:              []corev1.PersistentVolumeClaim{sampleClaim, sampleClaim2},
-				validateStorageClass: true,
-			},
-			wantErr: false,
-		},
-		{
-			name: "no claims: ok",
-			args: args{
-				k8sClient:            k8s.NewFakeClient(withVolumeExpansion(sampleStorageClass)),
-				initial:              nil,
-				updated:              nil,
-				validateStorageClass: true,
-			},
-			wantErr: false,
-		},
-		{
-			name: "claim in updated does not exist in initial: error",
-			args: args{
-				k8sClient:            k8s.NewFakeClient(withVolumeExpansion(sampleStorageClass)),
-				initial:              []corev1.PersistentVolumeClaim{sampleClaim},
-				updated:              []corev1.PersistentVolumeClaim{sampleClaim, sampleClaim2},
-				validateStorageClass: true,
-			},
-			wantErr: true,
-		},
-		{
-			name: "storage increase: ok",
-			args: args{
-				k8sClient:            k8s.NewFakeClient(withVolumeExpansion(sampleStorageClass)),
-				initial:              []corev1.PersistentVolumeClaim{sampleClaim, sampleClaim},
-				updated:              []corev1.PersistentVolumeClaim{sampleClaim, withStorageReq(sampleClaim, "3Gi")},
-				validateStorageClass: true,
-			},
-			wantErr: false,
-		},
-		{
-			name: "storage increase but volume expansion not supported: error",
-			args: args{
-				k8sClient:            k8s.NewFakeClient(&sampleStorageClass),
-				initial:              []corev1.PersistentVolumeClaim{sampleClaim, sampleClaim},
-				updated:              []corev1.PersistentVolumeClaim{sampleClaim, withStorageReq(sampleClaim, "3Gi")},
-				validateStorageClass: true,
-			},
-			wantErr: true,
-		},
-		{
-			name: "storage increase, volume expansion not supported, but no storage class check: ok",
-			args: args{
-				k8sClient:            k8s.NewFakeClient(&sampleStorageClass),
-				initial:              []corev1.PersistentVolumeClaim{sampleClaim, sampleClaim},
-				updated:              []corev1.PersistentVolumeClaim{sampleClaim, withStorageReq(sampleClaim, "3Gi")},
-				validateStorageClass: false,
-			},
-			wantErr: false,
-		},
-		{
-			name: "storage decrease: error",
-			args: args{
-				k8sClient:            k8s.NewFakeClient(withVolumeExpansion(sampleStorageClass)),
-				initial:              []corev1.PersistentVolumeClaim{sampleClaim, sampleClaim},
-				updated:              []corev1.PersistentVolumeClaim{sampleClaim, withStorageReq(sampleClaim, "0.5Gi")},
-				validateStorageClass: true,
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := ValidateClaimsStorageUpdate(context.Background(), tt.args.k8sClient, tt.args.initial, tt.args.updated, tt.args.validateStorageClass); (err != nil) != tt.wantErr {
-				t.Errorf("ValidateClaimsStorageUpdate() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
 func Test_validPVCModification(t *testing.T) {
 	es := func(nodeSets []esv1.NodeSet) esv1.Elasticsearch {
 		return esv1.Elasticsearch{
